% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmreduct.r
\name{asmreduction}
\alias{asmreduction}
\title{asmreduction conducts an age-structured stock reduction}
\usage{
asmreduction(inR0, fish, glb, props, limitH = c(0, 1), projyr = 0, constC = 0)
}
\arguments{
\item{inR0}{the trial value of unfished recruitment R0}

\item{fish}{a data.frame containing the year and catch in each year}

\item{glb}{the global variables defined in the data structures for
datalowSA}

\item{props}{the biological properties of the species, including
length-, weight-, maturity-, and selectivity-at-age}

\item{limitH}{a vecotr of two numbers denoting the lowest and
highest values of the maximum harvest rate the stock is
assumed to have experienced.}

\item{projyr}{number of years fo projecting at a constant catch. If
set to 0 the contents of constC are ignored}

\item{constC}{the constant catch to apply in the projections}
}
\value{
a list containing a summary matrix, and the full results
    for fully selected harvets rate, the spawning biomass, the
    depletion, and the explotiable biomass in each trajectory.
}
\description{
asmreduction conducts an age-structured stock
    reduction based on R functions out of the datalowSA package.
}
\examples{
\dontrun{
  data(orhdat1)
  fish <- orhdat1$fish # 19 years of data
  glb <- orhdat1$glb
  props <- orhdat1$props
  inR0 <- seq(12.0,14.2,0.02)
  limitH <- c(0.25,0.5)
  glb$M <- 0.032
  glb$steep <- 0.6
  reduct <- asmreduction(inR0,fish,glb,props,limitH=limitH)
  str(reduct,max.level = 1)
}
}
