% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{bh}
\alias{bh}
\title{bh calculates the expected Beverton-Holt recruitment}
\usage{
bh(spb, h, R0, B0)
}
\arguments{
\item{spb}{the current spawning or mature biomass}

\item{h}{the steepness of the Beverton-Holt stock recruitment curve}

\item{R0}{the unfished average recruitment level}

\item{B0}{the unfished spawning biomass.}
}
\value{
the expected Beverton-Holt recruitment level, a real number in the 
    linear scale
}
\description{
bh calculate the expected Beverton-Holt stock recruitment level
    from the available spawning biomass, the steepness, R0 and B0. This would
    be used when fitting a model to data.
}
\examples{
rec <- bh(10000,0.75,1500000,30000)
print(rec)   # should be 1285714
bh(spb=30000,h=0.75,R0=1500000,B0=30000)  # should be 1500000
}
