% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bhC}
\alias{bhC}
\title{bh calculates the expected Beverton-Holt recruitment}
\usage{
bhC(SPB, steep, R0, B0)
}
\arguments{
\item{SPB}{the current spawning or mature biomass}

\item{steep}{the steepness of the Beverton-Holt stock recruitment curve}

\item{R0}{the unfished average recruitment level}

\item{B0}{the unfished spawning biomass.}
}
\description{
bh calculate the expected Beverton-Holt stock recruitment 
    level from the available spawning biomass, the steepness, R0 and B0.
    This would be used when fitting a model to data.
}
