% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmreduct.r
\name{calcout}
\alias{calcout}
\title{calcout runs asmreduction across a vector of M values}
\usage{
calcout(mvalues, inR0, fish, glb, props, limitH = limitH)
}
\arguments{
\item{mvalues}{a vector of natural mortality values to replace those
inside the global object}

\item{inR0}{the trial value of log(R0)}

\item{fish}{a data.frame containing the year and catch in each year}

\item{glb}{the global variables defined in the data structures for
datalowSA}

\item{props}{the biological properties of the species, including
length-, weight-, maturity-, and selectivity-at-age}

\item{limitH}{a vector of two numbers denoting the lowest and
highest values of the maximum harvest rate the stock is
assumed to have experienced.}
}
\value{
a list of asmreduction lists
}
\description{
calcout runs asmreduction across an input vector of M 
    values and outputs a lisdt of aspreduction lists.
}
\examples{
\dontrun{
  data(orhdat1)
  fish <- orhdat1$fish # 19 years of data
  glb <- orhdat1$glb
  props <- orhdat1$props
  inR0 <- seq(12.0,14.2,0.02)
  limitH <- c(0.25,0.5)
  glb$M <- c(0.03,0.32,0.36,0.38,0.04,0.042,0.44)
  glb$steep <- 0.6
  out <- calcout(mvalues,inR0,fish,glb,props,limitH=limitH)
  str(out,max.level=1)
}
}
