% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{central}
\alias{central}
\title{central generates three estimates of central tendency}
\usage{
central(x, P = 0.9)
}
\arguments{
\item{x}{the vector of values whose central tendency is to be characertized}

\item{P}{the quantiles to be determined}
}
\value{
a 4 x 2 matrix containing the central tendency measures and the
   quantiles
}
\description{
central generates three estimates of central tendency and
    the quantiles about the distribution of the inut vector of values. The
    three measures are the arithmetic mean, the naive geometric mean, and
    the bias corrected geometric mean.
}
\examples{
\dontrun{
x <- rnorm(1000,mean=5,sd=1)
central(x,P=0.9)
x <- rlnorm(1000,meanlog=2,sdlog=0.2)
central(x,P=0.95)
exp(2)
}
}
