% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{checkdata}
\alias{checkdata}
\title{checkdata determines which methods match the input data}
\usage{
checkdata(indata, glob = NA)
}
\arguments{
\item{indata}{this can be either the complete list of data objects of just
the matrix of fisheries data.}

\item{glob}{in case one inputs a data.frame of fishery data this allows one
to enter the globals list containing as a minimum the spsname and
resilience, so that a test can be made for the spm analysis}
}
\value{
a 3 x 2 matrix with vaiable and true or false for presence
}
\description{
checkdata ensures the input fishery data contains the
    data necessary for each of the analyses present and reports which
    analyses are possible. It checks for simple errors
}
\examples{
\dontrun{
data(fishdat)
fish <- fishdat$fish
checkdata(fishdat)
checkdata(fish)
}
}
