% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchcurve.r
\name{classicCC}
\alias{classicCC}
\title{classicCC conducts a naive catch curve on the input data}
\usage{
classicCC(NM, ages, freqs, firstage = 0, plot = TRUE)
}
\arguments{
\item{NM}{an estimate of the natural mortality}

\item{ages}{the age classes encountered in the sample to be analysed}

\item{freqs}{the counts of each age-class in the sample}

\item{firstage}{the first age assumed to be fully selected by the fishing 
gear. younger ages are not included in the analysis.}

\item{plot}{a boolean determining whether to plot the catch curve, default
is TRUE}
}
\value{
a list containing the estimated fishing mortality, FF, the regression
    coefficients, and the model itself
}
\description{
classicCC conducts a classical catch curve in which one selects 
    the first age to be included in the calculation, assuming it is the 
    first age to be fully selected in the fishery. It merely applies a linear
    regression to the log-transformed counts by age and subtracts the natural
    mortality from teh gradient (total mortality) to obtain an estimate of 
    the fishing mortality experienced on average across the age-classes 
    included in the analysis.
}
\examples{
\dontrun{
age <- 0:9
counts <- c(3,4,501,2531,936,233,76,17,5,1)
M <- 0.6
pickage <- 3
out <- classicCC(M,age,counts,pickage,plot=TRUE)
print(out)
}           # NM=glb$M; ages=ages;freqs=freqs;firstage=7;plot=TRUE
}
