% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{cMSYphaseplot}
\alias{cMSYphaseplot}
\title{cMSYphaseplot plots the phase plot and catch and harvest rate plots}
\usage{
cMSYphaseplot(answer, fish, plotout = TRUE)
}
\arguments{
\item{answer}{the output from the run_cMSY function}

\item{fish}{the fishery data put into run_CMSY}

\item{plotout}{should a plot be produced, default=TRUE}
}
\value{
a list of meanB, meanH, msy, Bmsy, Hmsy, and Hlim, returned invisibly
}
\description{
cMSYphaseplot extracts the necessary data to enable the 
    production of a phase plot of estimated average biomass against estimated
    average harvets rate. It plots the Bmsy = 0.5B0 for the Schaefer model as 
    well as 20%B0 = 20%K as a default limit reference point. The first year 
    of data is identified by a larger green point and the last year by a 
    larger red point. It also plots the expected harvest rate that should 
    lead to MSY, called Ftarg and that, which if continued for long enough,
    would drive the biomass below the limit reference point, Flim. Points 
    above the Flim line (or Ftarg depending on which management objectives are
    used) would be classed as over-fishing leading to a status of 'depleting'
    and points to the left of 0.2B0 would be over-fished or 'depleted'.
    In addition, the function plots the catch history and the implied harvest
    rates just below the phase plot to aid in its interpretation.
}
\examples{
\dontrun{
  data(invert)
  fish <- invert$fish
  glb <- invert$glb
  answer <- run_cMSY(fish,glb,n=1000,sigpR=0.025)
  plotprep(width=7,height=6)
  cMSYphaseplot(answer,fish)
}
}
