% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{countgtzero}
\alias{countgtzero}
\title{countgtzero used in apply to count the number of zeros in a vector}
\usage{
countgtzero(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of zeros
}
\description{
countgtzero is designed to be used within the apply function to 
    count the number of zeros in a vector
}
\examples{
\dontrun{
x <- c(1,2,3,4,0,0,0,5,6,7,8) # 8 values > 0 
countgtzero(x)
}
}
