% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalowSA.r
\docType{package}
\name{datalowSA}
\alias{datalowSA}
\alias{datalowSA-package}
\title{datalowSA a set of functions to assist with data-poor assessments}
\description{
The datalowSA package provides three categories of functions
    analytical functions that enable the production of data-poor
    model-assisted stock assessments, utility functions that assist
    with data manipulation and extracting informaiton from output objects,
    and plotting functions that
    facilitate the illustration of the results of the assessments.
    In addition there are example data sets with which to illustrate the
    methods.
}
\section{Analytical functions}{

\describe{
  \item{central}{ generates three estimates of central tendency}
  \item{doproject}{ generates constant catch projections after running cMSY}
  \item{fillell}{ runs the checks on SRA to find success and failure}
  \item{fillell2}{ runs the checks on SRA to find success and failure, but
      but adds the criteria that the end depletion must be lower than the
      start}
  \item{Level4MSY}{ generates an MSY estimate from catches and F estimates}
  \item{run_cMSY}{ The main function for conducting a modified Catch-MSY
      analysis.}
  \item{sraMSY}{ Is called by run_cMSY and it calls oneSRA for as many
      iterations or replicates as entered. It produces large arrays of
      the biomass trajectories from each SRA along with whether or not
      each trajectory meets the acceptance criteria or not. Not exported
      but can be read using r4tier5:::sraMSY }
  \item{oneSRA}{ Is called by sraMSY. It takes in the vector of initial
      biomass depletions plus the randomly generated set of population
      model parameters and runs the SRA for each of the combinations of
      parameters and initial depletion levels. Not exported but can be
      read using r4tier5:::oneSRA }
  \item{pulloutStats}{ summaries the results from the Catch-MSY
      analysis by generating the mean, minimum, maximum, and quantiles of
      the resulting r, K, and MSY values.}
}
}

\section{Utility functions}{

\describe{
  \item{gettier5data}{ gets the columns of data required for Tier5, the
      input data.frame must contain at least year and catch, but can also
      contain species}
  \item{gettraject}{ extracts the plausible biomass trajectories from the
      output of cMSY}
  \item{halftable}{ halves the height of a tall narrow data.frame}
  \item{makedeplet}{ converts the biomass trajetories into a depletion matrix}
  \item{pulloutStats}{ summaries results from the Catch-MSY analysis}
  \item{datalowSA}{ A brief description of all functions in datalowSA}
  \item{summarycMSY}{ makes tables of msy, r, K, meanr, meanK, and all picks}
  \item{tier4to5}{ generates a Tier5 formatted dataset from a tier4 dataset}
  \item{whichsps}{ generates a listing of which species are in the tier4 data}
}
}

\section{Plotting functions}{

\describe{
  \item{plotMSY6}{ generates 6 graphs illustrating the array of rK
      parameter combinations and whether they were successful or not. That
      plot is coloured by how many trajectories across the initial
      depletion range were successful.}
  \item{plottrajectory}{ plots out the predicted biomass trajectories from
      those parameter combinations that have been accepted. It can either
      put all trajectories on one plot or generate a separate plot for each
      rK parameter set. Each individual biomass trajectory represents a set
      of population model parameters and a single initial depletion. It is
      possible to only print a specified number of parameter sets rather
      than all of them.}
}
}

\section{Data sets}{

\describe{
  \item{fishdat}{ A dataset containing the fish data.frame, the glb list, and the
    props data.frame set up ready for use with datalowSA. In particular it can
    be used with fitASPM, fitSPM, run_cMSY, and DBSRA. see ?fishdat}
  \item{dataspm}{ A dataset containing the fish data.frame, the glb list, and the
    props data.frame set up ready for use with datalowSA. In particular it can
    be used with the SPM functions, as well as the ASPM functions. see ?dataspm}
  \item{invert}{ A dataset containing the fish data.frame as a 31 x 7 matrix, 
    the glb and props data.frames are set to NULL. The fish data.frame has
    both the standardized cpue as well as the unstandardized geom, that is
    the geometric mean cpue.  This is particularly set up to
    be used with the SPM functions but also the Catch-MSY routines. see ?invert}
  \item{plaice}{ A dataset containing the fish, glb, props, agedata, and lendata
    for North sea plaice. Data taken from Beverton and Holt (1957). The primary 
    use of this data set is to illustrate the use of catch curves.} 
  \item{sps}{ A dataset containing 9 columns of typical scalefish fisheries data}   
}
}

\section{Vignettes}{

To learn more about datalowSA, start with the vignette:
\code{browseVignettes(package = "datalowSA")}
}

