% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{doconstC}
\alias{doconstC}
\title{doconstC calculates and plots projections under a constant catch}
\usage{
doconstC(inR1, projn = 5, constCatch = 100, lastyear = 2017,
  limit = 0.2, target = 0.48, console = TRUE, intensity = NA)
}
\arguments{
\item{inR1}{the R1 object that is within the list generated by run_cMSY.}

\item{projn}{the number of years of projection with a default of 5}

\item{constCatch}{the constant catch to be applied to each successful trajectory}

\item{lastyear}{the final year of the known catches and biomass trajectories}

\item{limit}{the depletion level acting as the limit referencepoint}

\item{target}{the depletion level used as a biomass target for the species.}

\item{console}{logical, should results be printed to the console. Default
=TRUE}

\item{intensity}{the value that defines the density of trajectories required
to give rise to full colour; default = NA which implies grey}
}
\value{
a matrix of the all years with the proportion < 20%, the 
    proportion > 48% (the input target), the mean and median depletion, and 
    the proportion of trajectories that were increasing relative to the 
    lastyear of data.
}
\description{
doconstC merely combines the functions gettraject, doproject,
    makedeplet, and plotconstC to simplify the process of conducting 
    projections using constant catches. Compare this with separately using 
    the functions gettraject, doproject, makedeplet, and plotconstC in 
    sequence, which is all that dococnstC does.
}
\examples{
\dontrun{
data(invert)
fish <- invert$fish
glb <- invert$glb
reps <- 10000  # one would run at least 20000, preferably more
answer <- run_cMSY(fish,glb,n=reps,sigpR=1e-6)
out <- doconstC(answer$R1,projn=5,constCatch=150,lastyear=2017,
                limit=0.2,target=0.4)
str(out)
}
}
