% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{doDepletion}
\alias{doDepletion}
\title{doDepletion - depletes the stock to the declared initdep level}
\usage{
doDepletion(inR0, indepl, inprops, inglb, inc = 0.02, Numyrs = 50)
}
\arguments{
\item{inR0}{the estimated log transformed unfished recruitment}

\item{indepl}{the initial depletion level to be searched for}

\item{inprops}{the props object from the data object}

\item{inglb}{the globals object from the data object}

\item{inc}{the starting value and step in harvest rate used when finding the 
selected depletion level; defaults to 0.02}

\item{Numyrs}{the number of years of fishing while searching for the
desired depletion. Some number between 40 - 50 seems to work well.}
}
\value{
A list containing the vector of numbers-at-age at the required 
    depletion level, the Fvalue required to achieve that depletion, and
    the actual depletion level achieved
}
\description{
doDepletion - depletes the stock to the declared initdep level
    There is a printout to the screen of the final outcome. The procedure
    assumes you have the unfished NaA in column 1 of the stock$NaA object.
    The depletion is arrived at by literally searching for the first
    constant harvest rate that leads to the required depletion.
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
props <- dataspm$props
dep <- doDepletion(glb$R0,indepl=0.4,props,glb,inc=0.02)
print(dep)
}  
}
