% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{doproject}
\alias{doproject}
\title{doproject after running the cMSY analysis the plausible trajectories}
\usage{
doproject(intraj, constC, projn = 5, sigpR = 0.025)
}
\arguments{
\item{intraj}{the successful biomass trajectories obtained from gettraject}

\item{constC}{the constant catch that is to be applied as a projection}

\item{projn}{the number of years to be projected.}

\item{sigpR}{the process error, the same as used in run_cMSY}
}
\value{
the same biomass trajecotry matrix as input excpet the empty years
    will have been filled with projections of the surplus production
    dynamics made under the constant catch level.
}
\description{
after running the run_cMSY analysis this function projects each
    of the accepted biomass trajectories
}
\examples{
\dontrun{
data(invert)
fish <- invert$fish
glb <- invert$glb
reps <- 5000  # one would run at least 20000, preferably more
answer <- run_cMSY(fish,glb,n=reps,sigpR=0.04)
traject <- gettraject(answer$R1,projn=5)
newtraj <- doproject(traject,constC=100)
head(newtraj)
trajdepl <- makedeplet(newtraj)
plotconstC(trajdepl,endyear=2017,constC=100,target=0.40)
}
}
