% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{dynamics}
\alias{dynamics}
\title{dynamics describe the ASPM dynamics}
\usage{
dynamics(pars, infish, inglb, inprops)
}
\arguments{
\item{pars}{the dynamics relies on many parameters sitting in the global
environment in particular ages, nages, maxage, M, maa, waa, sela, fish,
and nyrs. 'pars' can contain either two or three parameters. 1) is
the log-transformed average unfished recruitment, inR0. 2) is the 
variability around the index of relative abundance (cpue) during the 
fitting process, and if is present 3) is the initial depletion level 
initdepl, which if present will be fitted as well.}

\item{infish}{the fish data.frame from readdata or built in dataset}

\item{inglb}{the glb data.frame from readdata or built in dataset}

\item{inprops}{the props data.frame from readdata or built in dataset}
}
\value{
a data.frame containing the fishery dynamics according to the input
    parameter inR0. In particular it includes teh Catch and PredC, and the
    CPUE and PredCE, which can be used in a maximum likelihood context.
}
\description{
dynamics summarizes the dynamics of the Age-Structured
    Production Model (ASPM). Fitting the ASPM entails estimating the unfished
    recruitment level (R0), which is input as a parameter.
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
props <- dataspm$props
par <- c(glb$R0,0.20)  # not fitted to the data, this is just an initial guess
fishery <- dynamics(par,infish=fish,inglb=glb,inprops=props)
print(fishery)
}
}
