% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{ExB}
\alias{ExB}
\title{ExB calculate exploitable biomass from numbers-at-age}
\usage{
ExB(invect, SelA, WeightA)
}
\arguments{
\item{invect}{the numbers-at-age as a vector}

\item{SelA}{selectivity-at-age vector}

\item{WeightA}{weight-at-age vector as kilograms}
}
\value{
ExB a scalar as tonnes.
}
\description{
ExB calculates the spawning biomass from a vector of
    numbers-at-age, selectivity-at-age, and Weight-at-age.
}
\examples{
\dontrun{
data(fishdat)
str(fishdat)
glb <- fishdat$glb
fish <- fishdat$fish
props <- fishdat$props
unfish <- unfished(glb,props,glb$R0)
N1 <- unfish$N0 * exp(-glb$M/2)  # no growth
ExB(N1,props$sela,props$waa)  # should be 17999.6
}
}
