% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExBC}
\alias{ExBC}
\title{ExBC calculate exploitable biomass}
\usage{
ExBC(Nt, Sel, WghtA)
}
\arguments{
\item{Nt}{the numbers-at-age as a vector}

\item{Sel}{the selectivity-at-age vector}

\item{WghtA}{weight at age vector as kilograms}
}
\description{
ExBC calculates the exploitable biomass from a vector of
    numbers-at-age, selectivity-at-age, and Weight-at-age.
}
