% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fishNAAC}
\alias{fishNAAC}
\title{fishNAAC numbers-at-age dynamics using Rcpp.}
\usage{
fishNAAC(nyr, maxa, hS, hrate, glob, maa, waa, sela, Nt)
}
\arguments{
\item{nyr}{is literally the number of years of dynamics being run.}

\item{maxa}{is the maximum age}

\item{hS}{is half the natural survivorship exp(-M/2)}

\item{hrate}{is the fully selected annual harvest rate}

\item{glob}{is a vector of parameters of length at least 3:
index 0 = steepness, 1=B0, 2=R0,}

\item{maa}{is the maturity-at-age}

\item{waa}{is the weight-at-age}

\item{sela}{is the selectivity-at-age}

\item{Nt}{is thge numbers at age at time t}
}
\description{
conducts the number-at-age dynamics for use within the
    the age-structured production model.
}
