% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{getB0}
\alias{getB0}
\title{getB0 calculates the B0 from biological properties and R0}
\usage{
getB0(inR0, inglb, inprops)
}
\arguments{
\item{inR0}{the estimate of unfished recruitment}

\item{inglb}{the glb data.frame from readdata or built in dataset}

\item{inprops}{the props data.frame from readdata or built in dataset}
}
\value{
a single number that is the estimate of B0
}
\description{
getB0 calculates the B0 from biological properties of M,
    maxage, maa, and waa, plus the input of R0 on the nominal scale (the 
    hypothetical unfished recruitment level). This is used in the 'dynamics'
    function.
}
\examples{
\dontrun{
data(fishdat)
glb <- fishdat$glb
props <- fishdat$props
getB0(1275000,glb,props) # shoud give 19429.76
getB0(1000000,glb,props) # should give 15239.03
}
}
