% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{getLNCI}
\alias{getLNCI}
\title{getLNCI gets the log-normal confidence intervals}
\usage{
getLNCI(av, se, P = 95)
}
\arguments{
\item{av}{the mean value or a vector of mean values}

\item{se}{the standard deviation}

\item{P}{the percent used for the CI, defaults to 95.}
}
\value{
a vector of three for a single input or a matrix of 3 columns for 
    input vectors
}
\description{
getLNCI takes the mean and the standard deviation and produces
    the asymmetric log-normal confidence intervals around the mean values
}
\examples{
\dontrun{
  av <- c(4.0,2.15)
  se <- 0.33
  getLNCI(av,se,P=95)
  se <- c(0.33,0.4)
  getLNCI(av,se)
}
}
