% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{getprop}
\alias{getprop}
\title{getprop extract proportion of values between lim1 and lim2}
\usage{
getprop(invect, lim1 = 0, lim2 = 0)
}
\arguments{
\item{invect}{the collection of values to be subdivided}

\item{lim1}{the lower limit of values}

\item{lim2}{the upper limit of values}
}
\value{
a scalar containing the proportion of records
}
\description{
getprop extracts the proportion of values in invect either
    below lim1, between lim1 and lim2, or above lim2. If lim2 has the 
    value 0.0 and lim1 > 0 then the proportion below lim2 is extracted. 
    If lim1 = 0.0 and lim2 > 0, then the proportion > lim2 is extracted.
    If both lim1 and lim2 > 0 then the proportion between the two is
    extracted.
}
\examples{
\dontrun{
  x <- 1:100
  getprop(x,20)
  getprop(x,20,50)
  getprop(x,lim2=80)
}
}
