% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{greplow}
\alias{greplow}
\title{greplow - uses tolower in the search for the pattern}
\usage{
greplow(pattern, x)
}
\arguments{
\item{pattern}{- the text to search for in x}

\item{x}{- the vector or object within which to search for 'pattern' once
both have been converted to lowercase.}
}
\value{
the index location within x of 'pattern', if it is present, an empty
   integer if not
}
\description{
greplow - a grep implementation that ignores the case of either
   the search pattern or the object to be search. Both are converted to lower
   case before using grep.
}
\examples{
txt <- c("Long","Lat","LongE","LatE","Depth","Zone","Effort","Method")
greplow("zone",txt)
greplow("Zone",txt)
greplow("long",txt)
}
