% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalowSA.r
\docType{data}
\name{invert}
\alias{invert}
\title{invert data derived from a trawl caught invertebrate fishery.}
\format{
A list of three objects only two of which contains data
\describe{
  \item{fish}{ a data.frame containing year, catch, cpue, SE of the cpue, 
      geom, which is the unstandardized geometric mean CPUE, vessel, which
      is the number of active vessels reporting catches, and records, which is
      the number of cpue records reported each year }
  \item{glb}{ contains the resilience and spsname }
  \item{props}{ set to NULL}
}
}
\usage{
invert
}
\description{
A dataset containing the fish data.frame as a 31 x 7 matrix, 
    the glb and props data.frames are set to NULL. The fish data.frame has
    both the standardized cpue as well as the unstandardized geom, that is
    the geometric mean cpue.  This is particularly set up to
    be used with the SPM functions but also the Catch-MSY routines.
}
\examples{
 \dontrun{
 data(invert)
 str(invert)
 print(invert$fish)
}
}
\keyword{datasets}
