% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{logist}
\alias{logist}
\title{logist Logistic selectivity function}
\usage{
logist(inL50, delta, depend, knifeedge = 0)
}
\arguments{
\item{inL50}{is the length at 50 percent selection/maturity/whatever}

\item{delta}{is the difference in selection/maturity/whatever between
inL50 and inL95}

\item{depend}{a vector of lengths/ages for which the logistic value will be
calculated.}

\item{knifeedge}{defaults to 0. If knifeedge is set to a particular length or
age then the logistic value <= the value of knifeedge is set to
zero, which is essentially knife-edge. Allows for knife-edge selectivity}
}
\value{
A vector of length(depend) containing the predicted logistic values
}
\description{
logist calcualtes a Logistic curve that can be used as a
    selectivity function, or maturity curve, of wherever a logistic is
    required. This version uses the logistic function
    1/(1+exp(-log(19.0)*(lens-inL50)/(inL95-inL50))),
    which explicitly defines the SM50 and uses SM95 as the second parameter.
}
\examples{
\dontrun{
in50 <- 100.0
deltaS <- 8.0
lens <- seq(2,210,2)
select <- logist(inL50=in50,delta=deltaS,depend=lens)
selectk <- logist(in50,deltaS,lens,knifeedge=105)
round(cbind(lens[35:70],select[35:70],selectk[35:70]),5)
} 
}
