% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makebiomC}
\alias{makebiomC}
\title{makebiomC runs the surplus production model inside catch-MSY.}
\usage{
makebiomC(intheta, bd, ct)
}
\arguments{
\item{intheta}{the vector of parameters for the spm and the biomass}

\item{bd}{the intial biomass depletion vector}

\item{ct}{the vector of catches leading to the stock reduction}
}
\description{
runs the surplus production model inside catch-MSY. This is 
    written in C++ because it contains two for-loops the whole of which is
    run for however many iterations are used. Without using C++ the process
    can take 15 minutes rather than 20 seconds. This function is only called
    inside sraMSY, which is another function that did not strictly need to 
    be exported to the public gaze.
}
