% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{makecMSYdat}
\alias{makecMSYdat}
\title{makecMSYdat generates a cMSY dataset out of other data files}
\usage{
makecMSYdat(
  indat,
  spsname = "",
  yearcol = "Year",
  catchcol = "Catch",
  resil = "low"
)
}
\arguments{
\item{indat}{the tier 4 datafile for one species}

\item{spsname}{the name of the species to be analysed}

\item{yearcol}{the column name of the year data}

\item{catchcol}{the column namae of the total catch data}

\item{resil}{the value of resilience for the species as a character of
either "verylow", "low", "medium", or "high". Note the lower case,
defaults to "low"}
}
\value{
a list containing a matrix of year and catch, and a list of the
    resilience and spsname
}
\description{
makecMSYdat expects to find in the input
    data.frame or matrix, columns containing 'Year', and 'Total' 
    (meaning total removals = catch + discards), and 'Resilience' 
    being any of 'verylow', 'low', 'medium', or 'high'.
}
\examples{
\dontrun{
print("read the csv file containing the general data into 'dat'")
print("then run  makecMSYdat(dat,'spsname')")
}
}
