% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{makedeplet}
\alias{makedeplet}
\title{makedeplet converts the biomass trajectories into a depletion matrix}
\usage{
makedeplet(intraj)
}
\arguments{
\item{intraj}{the matrix derived from gettraject containing the successful
biomass trajectories.}
}
\value{
a matrix of only the biomass trajectories once they have each been
    divided by their respective K values
}
\description{
makedeplet converts the biomass trajectories into a deplletion
    matrix by dividing through each trajectory by its respective K value.
    Usually this would be done after the matrix had been projected forward.
}
\examples{
\dontrun{
traject <- rbind(c(rnorm(10,mean=200,sd=10),0.5,300,0.65),
                 c(rnorm(10,mean=200,sd=10),0.5,300,0.65))
colnames(traject) <- c(1:10,"r","K","initD")
makedeplet(traject)
}
}
