% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{multinomLL}
\alias{multinomLL}
\title{multinomLL is the negative log-likelihood for multinomial values}
\usage{
multinomLL(counts, predprop)
}
\arguments{
\item{counts}{the observed counts in each category (age or length) that is to
be fitted by the model}

\item{predprop}{the predicted proportions in each category. One needs a 
separate function to calculate this.}
}
\value{
a single scalar value, the -veLL
}
\description{
multinomLL calculates the negative log-likelihood for the 
    multinomial distribution. Commonly used with length and age composition 
    data in assessment models. Zero observations in particular categories
    do not matter as the equation is - Sum(Obs x log(expected-proportion)). 
    Thus zeros in the observations just contribute zero to the sum.
}
\examples{
\dontrun{
  obs <- c(0,5,11,19,12,3,0,0)
  expprop <- c(0.018,0.002,0.12,0.4,0.3,0.12,0.03,0.01)
  multinomLL(obs,expprop)  # should be 92.61393
}
}
