% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalowSA.r
\docType{data}
\name{orhdat1}
\alias{orhdat1}
\title{orhdat1 Three data objects suitable for use with asmreduct.}
\format{A list of three objects
\describe{
  \item{fish}{ a data.frame containing year, catch}
  \item{glb}{ a list of global variables including maxage, M, parameters 
      for growth, weight-at-age, maturity-at-age, steepness, R0, 
      selectivity, resilience, number of ages, the ages themselves, the
      number of years of catch data, and the species name}
  \item{props}{ a data.frame of age, laa, waa, maa, sela, and feca}
}}
\usage{
orhdat1
}
\description{
A dataset containing a fish data.frame, the glb list, and 
    the props data.frame set up ready for use with asmreduct.
}
\keyword{datasets}
