% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{outoptim}
\alias{outoptim}
\title{outoptim provides a sucinct summary of the optim solution}
\usage{
outoptim(inopt)
}
\arguments{
\item{inopt}{an object into which the output from optim is directed.}
}
\value{
nothing, but it does print the optim result to the console in five 
    lines without the list structure
}
\description{
outoptim provides a more succinct printout of the optim solution
    which only uses five lines on the screen rather than the standard print
    of the output list structure
}
\examples{
\dontrun{
data(fishdat)
bestL <- optim(c(14.0,0.3),aspmLL,method="Nelder-Mead",infish=fishdat$fish,
                inglb=fishdat$glb,inprops=fishdat$props,
                control=list(maxit=1000,parscale=c(10,0.1)))
outoptim(bestL)
}
}
