% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{penalty1}
\alias{penalty1}
\title{penalty1 enables the adding of a large penalty as one approaches 1.0}
\usage{
penalty1(x)
}
\arguments{
\item{x}{the parameter value that potentially incurs a penalty}
}
\value{
a single value as a penalty to be added to a Log-Likelihood or SSQ
}
\description{
penalty1 allows for the option of adding a large penalty as
    a parameter approaches 1.0 and moves to become larger than 1. See aspmLL 
    for example code that contains such a parameter. For example, when 
    fitting an ASPM sometimes the optimal mathematical model fit can occur 
    by depressing the R0 and having an initial depletion >>1. This is only
    used internally to datalowSA and so is not formally exported.
}
\examples{
datalowSA:::penalty1(0.5)
datalowSA:::penalty1(0.9)
datalowSA:::penalty1(0.99)
}
