% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{plotASPM}
\alias{plotASPM}
\title{plotASPM plots catch, CPUE, Spawning Biomass and Harvest Rate}
\usage{
plotASPM(infish, CI = NA, defineplot = TRUE, target = 0.48,
  usef = 7, png = "")
}
\arguments{
\item{infish}{an object generated by the dynamics function}

\item{CI}{defaults to NA, if confidence intervals around the cpue have been
obtained using getLNCI, then the resulting matrix will generate 95pc CIs}

\item{defineplot}{define the plot size and character outside the plot or
automatically inside. Defaults to TRUE}

\item{target}{target depletion level. Defaults to 0.48}

\item{usef}{defines the font to use usef(ont),default = 7 bold times}

\item{png}{save a png file with the name in 'png', default = "", which
means no file produced}
}
\value{
Nothing, but it does plot six graphs in a single plot.
}
\description{
plotASPM after running fitASPM the optimum parameters can be 
    put through the dynamics function to generate a dataframe containing
    the optimum dynamics. These can be plotted using plotASPM, which plots
    out the catches, the Spawning Biomass, the relative CPUE and its fit to
    the observed CPUE, and the harvest rate. This routine is still under 
    development to include more options.
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
props <- dataspm$props
pars <- c(14,0.3)
aspmLL(pars,fish,glb,props)      # should be -2.277029
bestspm <- fitASPM(pars,infish=fish,inglb=glb,inprops=props)
fishery <- dynamics(bestspm$par,fish,glb,props)
plotASPM(fishery,defineplot=TRUE)
ceCI <- getLNCI(fishery[,"PredCE"],bestspm$par[2])
plotASPM(fishery,CI=ceCI)
}  # infish=fishery; CI=NA; defineplot=TRUE; target=0.48; usef=7;png="test.png"
}
