% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{plotceASPM}
\alias{plotceASPM}
\title{plotceASPM plots just the fit of the ASPM model to the CPUE data}
\usage{
plotceASPM(infish, CI = NA, defineplot = TRUE)
}
\arguments{
\item{infish}{output from the dynamics function using the optimum parameters}

\item{CI}{the output matrix from getLNCI function. Needs to have the lower CI
in column 1 and the upper CI in the third column.}

\item{defineplot}{defaults to TRUE, determines whether to set up an new 
graphics window.}
}
\value{
returns nothing but does plot a graph
}
\description{
plotceASPM plots just the fit of the ASPM model to the CPUE data
    and provides a more detailed visual than plotASPM. It has the option of 
    including lognormal confidence intervals around the predcted CPUE.
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
props <- dataspm$props
pars <- c(14,0.3)
aspmLL(pars,fish,glb,props)      # should be -2.277029
bestspm <- fitASPM(pars,infish=fish,inglb=glb,inprops=props)
fishery <- dynamics(bestspm$par,fish,glb,props)
ceCI <- getLNCI(fishery[,"PredCE"],bestspm$par[2])
plotceASPM(fishery,CI=ceCI)
}
}
