% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{plotcMSY6}
\alias{plotcMSY6}
\title{plotcMSY6 plots out a summary of the Catch-MSY results in 6 graphs}
\usage{
plotcMSY6(cMSY, catch, label = NA)
}
\arguments{
\item{cMSY}{the list from running summcMSY on the output from run_cMSY}

\item{catch}{the catch in each year}

\item{label}{simply a text label for the y-axes; default = NA}
}
\value{
nothing, but it does generate a plot to the screen
}
\description{
plotcMSY6 generates 6 graphs illustrating the array of rK
    parameter combinations and whether they were successful or not. That
    plot is coloured by how many trajectories across the initial depletion
    range were successful.
}
\examples{
\dontrun{
data(invert)
fish <- invert$fish
glb <- invert$glb
reps <- 5000  # one would run at least 20000, preferably more
answer <- run_cMSY(fish,glb,n=reps,sigpR=0.04)
summcMSY <- summarycMSY(answer,fish,final=FALSE)
plotcMSY6(summcMSY,fish[,"catch"],label=glb$spsname)  
summcMSY <- summarycMSY(answer,fish,final=TRUE)
str(summcMSY,max.level=1)
plotcMSY6(summcMSY,fish[,"catch"],label=glb$spsname)
}
}
