% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{plotconstC}
\alias{plotconstC}
\title{plotconstC summarizes constant catch projections in catch-MSY}
\usage{
plotconstC(
  deplete,
  endyear,
  constC = 0,
  limit = 0.2,
  target = 0.48,
  console = TRUE,
  intensity = NA,
  contours = TRUE
)
}
\arguments{
\item{deplete}{the output of 'gettraject', a matrix of biomass trajectories}

\item{endyear}{the final year of the known catches and biomass trajectories}

\item{constC}{the constant catch applied to generate the proections}

\item{limit}{biomass depletion level used as a limit for the species}

\item{target}{the depletion level used as a target for the species concerned.}

\item{console}{a boolean determinng whether the projection years results are
printed to the console}

\item{intensity}{defaults to NA but if contains a value this is the density
of trajectories that lead to full colour}

\item{contours}{default to TRUE, should the 80% and 90% quantile contours 
be plotted?}
}
\value{
a matrix of all years with the proportion < 20%, the proportion > 48% 
    (the input target), the mean and median depletion, and the proportion of 
    trajectories that were increasing relative to the endyear of data.
}
\description{
plotconstC plots and summarizes the outcome of constant 
    catch projections made on each of the successful trajectories from the 
    catch-MSY analysis. The catch-MSY analysis provides an uncertain estimate 
    of MSY and of final depletion. But by conducting constant catch
    projectios the implications in terms of what proportion of trajectories 
    increase and what proportion decrease can aid decisions and could form 
    he basis for the development of formal harvest control rules. It will be
    noticed that a large proportion of trajectories can be below 20%, but
    notice also that at the same time a high proportion can be above 48%, 
    assuming that this Commonwealth default target is used. It is the case 
    that a target of 0.4 or 40% has been suggested for non-primary target
    species, as a means of preventing the management of such species from 
    prevent the primary economic drivers of a fishery from being caught.
}
\examples{
\dontrun{
data(invert)
fish <- invert$fish
glb <- invert$glb
reps <- 10000  # one would run at least 20000, preferably more
answer <- run_cMSY(fish,glb,n=reps,sigpR=1e-8)
traject <- gettraject(answer$R1,projn=5)
newtraj <- doproject(traject,constC=300)
trajdepl <- makedeplet(newtraj)
plotconstC(trajdepl,endyear=2017,constC=300,limit=0.2,target=0.40)
}
}
