% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmreduct.r
\name{plotdepletion}
\alias{plotdepletion}
\title{plotdepletion}
\usage{
plotdepletion(inreduct, defineplot = TRUE, refpts = c(0.2, 0.48))
}
\arguments{
\item{inreduct}{the list object generated by asmreduction,
containing the result of the age-structured stock reduction}

\item{defineplot}{a boolean determining whether to define the pars
for a plot or not. defaults to TRUE}

\item{refpts}{a vector of two defining the limit and target
reference points; defaults to 0.2 and 0.48}
}
\value{
nothing but it does generate a plot
}
\description{
plotdepletion
}
\examples{
\dontrun{
  data(orhdat1)
  fish <- orhdat1$fish # 19 years of data
  glb <- orhdat1$glb
  props <- orhdat1$props
  inR0 <- seq(12.0,14.2,0.02)
  limitH <- c(0.25,0.5)
  glb$M <- 0.032
  glb$steep <- 0.6
  reduct <- asmreduction(inR0,fish,glb,props,limitH=limitH)
  plotprep(width=6, height=4)
  plotdeletion(reduct)
}
}
