% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{plotfish}
\alias{plotfish}
\title{plotfish plots the catch and optionally the cpue from fish}
\usage{
plotfish(fish, glb, ce = TRUE, title = TRUE, fnt = 7,
  filename = "", resol = 200)
}
\arguments{
\item{fish}{the matrix or data.frame containing year, catch, and cpue.}

\item{glb}{the list of biologicals potentially containing the spsname}

\item{ce}{a logical parameter determining whether to plot the cpue or not.
the default = TRUE}

\item{title}{determines whether or not the spsname is printed at the top
of the plot. Default = TRUE but for a more formal publication it 
might need to be set to FALSE, which also reallocates the room 
given to the title to the plot.}

\item{fnt}{the font used in the plot and axes.}

\item{filename}{default is empty. If a filename is put here a .png file
with that name will be put into the working directory.}

\item{resol}{the resolution of the png file, defaults to 200 dpi}
}
\value{
prints the location of the png file produced to the console
}
\description{
plotfish uses the matrix of fishery data used in the
    datalowSA standard data format. It requires the matrix or data.frame
    to contain the columns 'year', 'catch', and optionally 'cpue'.
}
\examples{
\dontrun{
  data(dataspm)
  plotfish(dataspm$fish,glb,ce=TRUE,filename="")
}
}
