% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmreduct.r
\name{plotout}
\alias{plotout}
\title{plotout}
\usage{
plotout(out)
}
\arguments{
\item{out}{a list of asmreduction lists generated by calcout}
}
\value{
a matrix of logR0, DeplHH, logR0, DeplLH, and Steepness
}
\description{
plotout
}
\examples{
\dontrun{
  data(orhdat1)
  fish <- orhdat1$fish # 19 years of data
  glb <- orhdat1$glb
  props <- orhdat1$props
  inR0 <- seq(12.0,14.2,0.02)
  limitH <- c(0.25,0.5)
  glb$M <- c(0.03,0.32,0.36,0.38,0.04,0.042,0.44)
  glb$steep <- 0.6
  out <- calcout(mvalues,inR0,fish,glb,props,limitH=limitH)
  plotprep(width=6, height=4)
  plotout(out)
}
}
