% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmreduct.r
\name{plotreduction}
\alias{plotreduction}
\title{plotreduction generates a summary plot of a stock reduction}
\usage{
plotreduction(inreduct, defineplot = TRUE, refpts = c(0.2, 0.48))
}
\arguments{
\item{inreduct}{the list object generates by asmreduction}

\item{defineplot}{boolean which determines whether a par statement
is made or not. default = TRUE.}

\item{refpts}{a vector of two defining the limit and target
reference points; defaults to 0.2 and 0.48}
}
\value{
outputs the stats from any projection period. also produces
    a 3,1 plot of FullH, spawning biomass, and depletion for the
    input stock reduction
}
\description{
plotreduction generates a summary plot of the output
    from an age-structured stock reduction produced by the
    asmreduction function, which in turn relies on the dynamics
    function from the aspm within the datalowSA package.
}
\examples{
\dontrun{
  data(orhdat1)
  fish <- orhdat1$fish # 19 years of data
  glb <- orhdat1$glb
  props <- orhdat1$props
  inR0 <- seq(12.0,14.2,0.02)
  limitH <- c(0.25,0.5)
  glb$M <- 0.032
  glb$steep <- 0.6
  reduct <- asmreduction(inR0,fish,glb,props,limitH=limitH)
  plotprep(width=6,height=5.5)
  plotreduction(reduct)
}     # inreduct=reduct; defineplot=TRUE
}
