% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{plottrajectory}
\alias{plottrajectory}
\title{plottrajectory literally plots predicted trajectories}
\usage{
plottrajectory(
  inR1,
  years,
  catch,
  inparbound,
  scalar = 1000,
  Bmax = 2,
  oneplot = TRUE,
  plotout = TRUE,
  plotall = 7
)
}
\arguments{
\item{inR1}{the output from run_cMST}

\item{years}{the vector of years}

\item{catch}{the vector of catches per year}

\item{inparbound}{the set of parameter vectors that were successful}

\item{scalar}{literally scales the catch to the same units as biomass;
defaults to 1000 so as to convert Kg to tonnes.}

\item{Bmax}{Deprecated. No longer used. Remove from any code will be depleted
from later iterations of datalowSA}

\item{oneplot}{Plot all trajectories on top of each other rather than
individually. defaults to TRUE.}

\item{plotout}{produce a plot or not? Defaults to TRUE}

\item{plotall}{when oneplot=FALSE how many plots to generate; defaults to
7, which plots the successful trajectories from the first seven r-K pairs
that had successful outcomes. Useful numbers are 7 and 15 as the total
catch history is also illustrated along with the mean MSY to aid in 
understandig the trajectories. To see all trajectories set plotall=TRUE.}
}
\value{
a list of the yes/no vector and of the accepted rK pairs. This is
    returned invisibly.
}
\description{
plottrajectory plots out the predicted trajectories from
    those paramter combinations that have been accepted. In addition, and
    more importantly, it identifies those trajectories that succeeded and
    puts them into a smaller matrix than the complete set of trialed rK
    combinations and each of the successful trajectories. This can project 
    a limited number of trajectories, determined by oneplot and plotall. If
    oneplot is true it does not matter what is in plotall, all trajectories
    are given in a single plot along with the median values in red. Similarly 
    for the harvest rate
}
\examples{
\dontrun{
 data(invert)
 fish <- invert$fish 
 glb <- invert$glb 
 reps <- 5000  # one would run at least 20000, preferably more 
 answer <- run_cMSY(fish,glb,n=reps,sigpR=0.04) 
 summcMSY <- summarycMSY(answer,fish,final=TRUE) 
 # plotprep(width=8,height=5,newdev=FALSE)
 out <- plottrajectory(answer$R1,fish$year,fish$catch,answer$parbound,
                      oneplot=FALSE,scalar=1.0,plotout=TRUE,plotall=7)
}  #
}
