% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchcurve.r
\name{popN}
\alias{popN}
\title{popN calculates the proportional catch at age for catch curves}
\usage{
popN(maxa, M, initF, select)
}
\arguments{
\item{maxa}{the maximum age in the population and sample}

\item{M}{the natural mortality estimate}

\item{initF}{the fishing mortality to be tested}

\item{select}{the selectivity of the fishery}
}
\value{
a vector of the expected proportional distribution of numbers-at-age
}
\description{
popN calculates the expected proportional catch-at-age for use
    when using a multinomial likeihood to fit a catch curve that has been
    enhanced through the addition of a logistic selectivity curve. It
    generates the expected equilibrium relative numbers-at-age in the 
    population and then uses the Baranov catch equation to generate the 
    expected numbers-at-age in the catch, and hence to be expected in the
    catch curve. This approach requires three parameters while the classicCC
    uses only two parameters. This function is 
    used within the function selectCC, that fits a catch curve enhanced with
    a selectivity curve so as to use the counts from all ages. It can also
    be used when plotting the outcome of a catch curve analysis.
}
\examples{
\dontrun{
M <-  0.6
age <- 0:9
counts <- c(3,4,501,2531,936,233,76,17,5,1)
pars <- c(lm50=2.5,delta=0.5,fcur=0.6)
pp <- popN(max(age),M,pars[3],logist(pars[1],pars[2],age))
round(cbind(age,counts/sum(counts),pp),5)
}
}
