% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{pulloutStats}
\alias{pulloutStats}
\title{pulloutStats summaries results from the Catch-MSY analysis}
\usage{
pulloutStats(inR1, probabs = c(0.025, 0.05, 0.5, 0.95, 0.975))
}
\arguments{
\item{inR1}{the input parameter vectors with their respective ok values}

\item{probabs}{the percentiles used in pulling out the quantiles of the
r, K, and MSY values; default c(0.025, 0.05, 0.1,0.5,0.9, 0.95, 0.975)}
}
\value{
a list containing a matrix of summary statistics regarding the r, K, 
    MSY, and last year depletion values, he biomass trajectories, and the 
    depletion trajectories
}
\description{
pulloutStats summaries the results from the Catch-MSY
    analysis by generating the mean, minimum, maximum, and quantiles of
    the resulting r, K, MSY, and last year depletion values.
}
\examples{
\dontrun{
data(invert)
fish <- invert$fish
glb <- invert$glb
nyr <- length(fish$year)
answer <- run_cMSY(fish,glb,n=5000,sigpR=0.04)
results <- pulloutStats(answer$R1)
str(results)
}           
}
