% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchcurve.r
\name{selectCC}
\alias{selectCC}
\title{selectCC conducts a selectivity enhanced catch curve analysis}
\usage{
selectCC(M, maxage, counts, pars, plot = FALSE)
}
\arguments{
\item{M}{the natural mortality estimate}

\item{maxage}{is the maximum age in the available data}

\item{counts}{is the numbers at age, a vector of counts with names that are the
ages to which the counts relate}

\item{pars}{three values in a vector, these are the initial estimates of the 
two selectivity parameters (A50, the age at 50% selection, and delta = 
(A95 - A50), of course delta must be positive), and the initial 
estimate (guess) of the fishing mortality. See the example below.}

\item{plot}{a logical parameter determining whether to plot the result or not,
the default is FALSE}
}
\value{
a list containing the output from optim, 'best', with the parameter 
    estimates and the other diagnostics from the fit, and a matrix, 'result' 
    containing the observed and predicted values and counts
}
\description{
selectCC encapsulates the routines required to fit a selectivity
    enhanced catch curve to a catch-at-age composition sample. The method is 
    described in Wayte, S.E. and N.L. Klaer (2010) An effective harvest 
    strategy using improved catch-curves. Fisheries Research 106: 310-320. 
    Although the log-likelihood used here is based on the strict multinomial
    distribution.
}
\examples{
\dontrun{
M <-  0.6
ages <- 0:9
counts <- c(3,4,501,2531,936,233,76,17,5,1)
names(counts) <- 0:9
pars <- c(A50=2.5,delta=0.5,fcur=0.6)
selectCC(M,max(ages),counts,pars)
# now with a plot
selectCC(M,max(ages),counts,pars,plot=TRUE)
}  
}
