% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{simpfox}
\alias{simpfox}
\title{simpfox simply calculates the predicted CE for an SPM; deprecated}
\usage{
simpfox(par, indat)
}
\arguments{
\item{par}{the parameters of the SPM = r, K, and Binit}

\item{indat}{the data which needs to include year, catch, and CPUE}
}
\value{
a vector of nyrs of the predicted CPUE
}
\description{
simpfox calculates the predicted CPUE for a Fox SPM model. This
    function is now deprecated and instead a schaefer parameter (TRUE or 
    FALSE) has been introduced into simpspm
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
fish
colnames(fish) <- tolower(colnames(fish)) 
pars <- c(r=0.138,K=6129,Binit=2757) 
predCE <- simpfox(pars,fish)
cbind(fish[,"year"],fish[,"cpue"],predCE)
print("")
print("Deprecated, use simpspm and schaefer=FALSE instead.")
}
}
