% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{simpspm}
\alias{simpspm}
\title{simpspm simply calculates the predicted CE for an SPM}
\usage{
simpspm(par, indat, schaefer = TRUE)
}
\arguments{
\item{par}{the parameters of the SPM = r, K, and Binit}

\item{indat}{the data which needs to include year, catch, and CPUE}

\item{schaefer}{a logical value determining whether the spm is to be a
simple Schaefer model (p=1) or approximately a Fox model (p=1e-08). The
default is TRUE}
}
\value{
a vector of nyrs of the predicted CPUE
}
\description{
simpspm calculates the predicted CPUE for an SPM model. It
    assumes that there is a variable called 'p' in the global environment
    and this 'p' variable determines the asymmetry of the production curve.
    If p = 1.0 then the SPM is the Schaefer model, if it is 1e-8 it
    approximates the Fox model.
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
fish
colnames(fish) <- tolower(colnames(fish)) 
pars <- c(r=0.242,K=5170,Binit=2840) 
predCE <- simpspm(pars,fish) 
cbind(fish[,"year"],fish[,"cpue"],predCE)
}
}
