% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{SpB}
\alias{SpB}
\title{SpB - calculate spawning biomass from a vector of numbers-at-age}
\usage{
SpB(invect, MatureA, WeightA)
}
\arguments{
\item{invect}{- the numbers-at-age as a vector}

\item{MatureA}{maturity at age vector}

\item{WeightA}{weight at age vector as kilograms}
}
\value{
SpB - a scalar as tonnes.
}
\description{
SpB - calculates the spawning biomass from a vector of
    numbers-at-age, Maturity-at-age, and Weight-at-age.
}
\examples{
\dontrun{
data(fishdat)
str(fishdat)
glb <- fishdat$glb
fish <- fishdat$fish
props <- fishdat$props
unfish <- unfished(glb,props,glb$R0)
N0 <- unfish$N0
SpB(N0,props$maa,props$waa)  # should be 18326.52
unfish$B0
}
}
