% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SpBC}
\alias{SpBC}
\title{SpBC calculate spawning biomass from a vector of numbers-at-age}
\usage{
SpBC(Nt, MatA, WghtA)
}
\arguments{
\item{Nt}{the numbers-at-age as a vector}

\item{MatA}{maturity at age vector}

\item{WghtA}{weight at age vector as kilograms}
}
\description{
SpBC calculates the spawning biomass from a vector of
    numbers-at-age, Maturity-at-age, and Weight-at-age.
}
