% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{sraMSY}
\alias{sraMSY}
\title{sraMSY sets up the input parameters and data for oneSRA}
\usage{
sraMSY(theta, N, startbd, nyr, ct, yr, mult, maxH, Fyear)
}
\arguments{
\item{theta}{a list containing the initial parameter values including
the initial r range (2 numbers), the initial k range (2 numbers), the
final depletion range (2 numbers), and the recruitment variability}

\item{N}{the number of replicate searches across the parameter space}

\item{startbd}{the set of initial starting depletion levels}

\item{nyr}{the number of years of catch data}

\item{ct}{the vector of catches per year}

\item{yr}{the vector of years}

\item{mult}{a multiplier for K to allow for stock biomasses to rise above K;
default value = 1.0, so K is the default upper limit}

\item{maxH}{upper limit of harvest rate included in the constraints;
defaults to 1.0, which implies no upper limit.}

\item{Fyear}{is the index to the year in which a rnage of harvest rates
is to be used to constrina the acceptable trajectories.}
}
\value{
a list containing itheta (the vectors of parameters), elltot (a
    matrix of the yes/no vectors for each initial biomass depletion level),
    and biomass (the biomass trajectories for each of the N parameter
    vectors)  # theta=parbound;N=n;startbd=startbd;nyr=nyr;ct=ct;yr=yr;mult=multK; maxH=1.0;Fyear=NA
}
\description{
sraMSY sets up the run parameters in terms of input parameter
    bounds, the number of replicates, and other parameters.
    Not exported but can be viewed using datalowSA:::sraMSY.
}
