% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{summarycMSY}
\alias{summarycMSY}
\title{summarycMSY makes tables of msy,r,K,meanr,meanK,and all picks}
\usage{
summarycMSY(ans, fish, final = TRUE)
}
\arguments{
\item{ans}{the object output from run_cMSY}

\item{fish}{the input data often named fish}

\item{final}{whether or not to consider the first phase or the final phase,
the default is TRUE, which considers the final phase}
}
\value{
a list of 12 objects used to summarize and plot the output. Returned
    invisibly; need to first allocate to an object.
}
\description{
summarycMSY generates a list of countcolour, meanmsy, meanr,
    meanK, r, K, msy, pickC (a list of pickblk, pickblu, pickyel,pickmax,
    rnot, knot), and years
}
\examples{
\dontrun{
data(invert)
fish <- invert$fish
glb <- invert$glb

nyr <- length(fish$year)
reps <- 5000  # one would run at least 20000, preferably more
answer <- run_cMSY(fish,glb,n=reps,sigpR=0.04)
summcMSY <- summarycMSY(answer,fish,final=TRUE)
str(summcMSY,max.level=1)
}
}
