% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalow_utils.r
\name{toXL}
\alias{toXL}
\title{toXL copies a data.frame or matrix to the clipboard}
\usage{
toXL(x, output = FALSE)
}
\arguments{
\item{x}{a vector or matrix}

\item{output}{- a boolean determining whether to print the object to the
screen as well as the clipboard; defaults to FALSE}
}
\value{
Places the object 'x' into the clipboard ready for pasting
}
\description{
toXL copies a data.frame or matrix to the clipboard
   so one can then switch to Excel and just type <ctrl> + V to paste the
   data in place
}
\examples{
\dontrun{
datamatrix <- matrix(data=rnorm(100),nrow=10,ncol=10)
colnames(datamatrix) <- paste0("A",1:10)
rownames(datamatrix) <- paste0("B",1:10)
toXL(datamatrix,output=TRUE)
}
}
