% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMSY_funs.R
\name{trendMSY}
\alias{trendMSY}
\title{trendMSY calculates the  mean MSY per K class}
\usage{
trendMSY(inr, inK, inc = 100)
}
\arguments{
\item{inr}{the set of r values to be tested, derives from summarycMSY}

\item{inK}{the set of K values to be tested, derives from summarycMSY}

\item{inc}{the class width of the K classes}
}
\value{
a matrix of r, K and MSY values
}
\description{
trendMSY subdivides the range of the successful K values
    into a set of classes of width 'inc' and for each class calculates the
    mean (geometric mean) of the MSY for the subset of r and K values. This
    enables the central value of MSY to be plotted on the scatter of 
    successful points.
}
\examples{
\dontrun{
 data(invert)
 fish <- invert$fish 
 glb <- invert$glb 
 reps <- 5000  # one would run at least 20000, preferably more 
 answer <- run_cMSY(fish,glb,n=reps,sigpR=0.04) 
 summcMSY <- summarycMSY(answer,fish,final=TRUE) 
 trendMSY(summcMSY$r,summcMSY$K,inc=200) 
}
}
