% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspm.r
\name{unfished}
\alias{unfished}
\title{unfished generates the numbers at age for an unfished population}
\usage{
unfished(glob, propert, inR0)
}
\arguments{
\item{glob}{the global constants object containing biology and structure}

\item{propert}{the data.frame containing laa, maa, waa, maa, and sela}

\item{inR0}{the unfished recruitment from B0}
}
\value{
a list containing N0, ExNO, R0, A0, B0, and ExB0
}
\description{
unfished generates the numbers at age for an unfished
    population, and determines the recruitment dynamics. It requires the
    input of R0. The output includes the unfished numbers-at-age N0, and the 
    unfished exploitable biomass, which is from the numbers-at-age after 
    half of natural mortality, ExN0.
}
\examples{
\dontrun{
data(fishdat)
glb <- fishdat$glb
fish <- fishdat$fish
props <- fishdat$props
unfish <- unfished(glb,props,glb$R0)
print(unfish)
}
}
